add models and refactor code
