+++ app/src/main/java/com/picpay/desafio/android/data/remote/data_source/UserRDS.kt
@@ -1,10 +1,11 @@
 package com.picpay.desafio.android.data.remote.data_source
 
-import com.picpay.desafio.android.data.remote.model.User
+import com.picpay.desafio.android.data.remote.model.UserRM
+import com.picpay.desafio.android.domain.model.User
 import retrofit2.Call
 import retrofit2.http.GET
 
-// User Remote Data source
+// UserRM Remote Data source
 interface UserRDS {
 
     @GET("users")
