+++ app/src/main/java/com/picpay/desafio/android/domain/model/User.kt
@@ -1,0 +1,8 @@
+package com.picpay.desafio.android.domain.model
+
+data class User(
+        val img: String,
+        val name: String,
+        val id: Int,
+        val username: String
+    )
