+++ app/src/test/java/com/picpay/desafio/android/ExampleService.kt
@@ -1,7 +1,8 @@
 package com.picpay.desafio.android
 
 import com.picpay.desafio.android.data.remote.data_source.UserRDS
-import com.picpay.desafio.android.data.remote.model.User
+import com.picpay.desafio.android.data.remote.model.UserRM
+import com.picpay.desafio.android.domain.model.User
 
 class ExampleService(
     private val service: UserRDS
